Rename chart resource to release
