+++ service/controller/chart/v1/resource/release/error.go
@@ -1,4 +1,4 @@
-package chart
+package release
 
 import "github.com/giantswarm/microerror"
